[Refactor] Move workflow restore from app to GraphCanvas
